Refactor to make CreateEventPage the parent component
