+++ src/components/App.js
@@ -19,7 +19,7 @@ import PremiumF2FMembershipPage from '../containers/PremiumF2FMembershipPage'
 import GettingStartedPage from '../containers/GettingStartedPage'
 import EventsList from '../containers/EventsList'
 import EventInfo from '../containers/EventInfo'
-import EventForm from '../containers/EventForm'
+import CreateEventPage from '../containers/CreateEventPage'
 import { withCookies } from 'react-cookie'
 
 class App extends Component {
@@ -77,7 +77,7 @@ class App extends Component {
           <Route exact path='/events' component={EventsList} />
           <Route path='/events/new' render={props => {
             return (
-              <EventForm
+              <CreateEventPage
                 {...props}
                 cookies={this.props.cookies}
               />)
