--- src/components/CreateEventPage.js
@@ -1,12 +1,0 @@
-import React from 'react'
-import { Header } from 'semantic-ui-react'
-
-const CreateEventPage = () => {
-  return (
-    <Header as='h1' textAlign='center'>
-          Creating a new Event
-    </Header>
-  )
-}
-
-export default CreateEventPage
